+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Update server to use `TickPolicy` instead of requiring a tick rate.
+
 ## [0.4.0] - 2023-05-26
 
 ### Changed